requirements.txt
troposphere
awacs
boto3
psycopg2-binary
simple-crypt
